+++ src/react/components/modal/__tests__/XUIModal-test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, { mount, shallow } from 'enzyme';
+import renderer from 'react-test-renderer';
 import Adapter from 'enzyme-adapter-react-15';
-import toJson from 'enzyme-to-json';
 import XUIModal from '../XUIModal';
 import XUIModalBody from '../XUIModalBody';
 import XUIModalFooter from '../XUIModalFooter';
@@ -9,19 +9,22 @@ import XUIModalHeader from '../XUIModalHeader';
 
 Enzyme.configure({ adapter: new Adapter() });
 
+const NOOP = () => {};
+
 // NOTE: The esc key functionality has not been tested here as event listeners don't work very well in React testing
 describe('XUIModal', () => {
 	it('Should be possible to do snapshot testing on the modal', () => {
-		const component = shallow(
+		const component = renderer.create(
 			<XUIModal
 				isUsingPortal={false}
 				id="test-modal"
-				onClose={() => {}}
+				qaHook="xui-modal"
+				onClose={NOOP}
 			>
 				<div>test</div>
 			</XUIModal>
 		);
-		expect(toJson(component)).toMatchSnapshot();
+		expect(component).toMatchSnapshot();
 	});
 
 	it('Should render with any additional classes provided through the className, maskClassName and closeClassName props', function () {
@@ -94,12 +97,12 @@ describe('XUIModal', () => {
 		expect(modalFullScreen.hasClass('xui-modal-fullscreen')).toBeTruthy();
 	});
 
-	it('Should correctly pass close handlers to close button and mask', function () {
+	it('Should correctly pass close handlers to the close button and mask', function () {
 		let clicked = false;
-		const demoFunction = ()=>{clicked = true;};
+		const demoFunction = ()=>{clicked = true};
 
 		const component = mount(
-			<XUIModal isUsingPortal={false} onClose={demoFunction} isHidden={false} hideOnOverlayClick={true}>
+			<XUIModal isUsingPortal={false} onClose={demoFunction} isOpen={true} hideOnOverlayClick={true}>
 			</XUIModal>
 		);
 		const maskNode = component.find('.xui-mask');
@@ -122,20 +125,42 @@ describe('XUIModal', () => {
 				isUsingPortal={false}
 				id='test-modal-overlayClick'
 				onClose={demoFunction}
-				isHidden={false}
 				hideOnOverlayClick={false}
+				isOpen={true}
 			>
 			</XUIModal>
 		);
 		const maskNode = component.find('.xui-mask');
 		maskNode.simulate('click');
-		expect(maskClicked).toEqual(false);
+		setTimeout(()=> {
+			expect(maskClicked).toBeFalsy();
+		}, 0);
 	});
 
-	it('Should only add default layout classes if defaultLayout isn\'t set to false', function () {
+	it('Should fire on close on click of the mask by default', () => {
+		let maskClicked = false;
+		const demoFunction = () => { maskClicked = true; };
+
+		const component = mount(
+			<XUIModal
+				isUsingPortal={false}
+				id='test-modal-overlayClick'
+				onClose={demoFunction}
+				isOpen={true}
+			>
+			</XUIModal>
+		);
+		const maskNode = component.find('.xui-mask');
+		maskNode.simulate('click');
+		setTimeout(()=> {
+			expect(maskClicked).toBeTruthy()
+		}, 0);
+	});
+
+	it('Should only add default layout classes if defaultLayout is explicitly set to true or by default', function () {
 		const component = mount(
 			<div>
-				<XUIModal isUsingPortal={false} id="test-modal-nolayout" defaultLayout={false}>
+				<XUIModal isUsingPortal={false} id="test-modal-nolayout" defaultLayout={false} isOpen={true}>
 				</XUIModal>
 				<XUIModal isUsingPortal={false} id="test-modal-layout">
 				</XUIModal>
@@ -150,7 +175,7 @@ describe('XUIModal', () => {
 
 	it('Should not add margin or padding classes to subcomponents if defaultLayout is set to false', function () {
 		const component = mount(
-			<XUIModal isUsingPortal={false} defaultLayout={false} id="test-modal-layoutSecond">
+			<XUIModal isUsingPortal={false} defaultLayout={false} id="test-modal-layoutSecond" isOpen={true}>
 				<XUIModalHeader defaultLayout={false}/>
 				<XUIModalBody defaultLayout={false}/>
 				<XUIModalFooter defaultLayout={false}/>
@@ -159,15 +184,13 @@ describe('XUIModal', () => {
 		const modal = component.find('#test-modal-layoutSecond .xui-modal');
 
 		for (let i=0; i < modal.children.length; i++) {
-			expect(modal.childAt(i).hasClass('xui-margin')).toBeFalsy();
-			expect(modal.childAt(i).hasClass('xui-padding')).toBeFalsy();
 			expect(modal.childAt(i).hasClass('xui-modal-layout')).toBeFalsy();
 		}
 	});
 
 	it('Should render as a form when isForm is true', function () {
 		const component = mount(
-			<XUIModal isUsingPortal={false} isForm={true} id="test-modal">
+			<XUIModal isUsingPortal={false} isForm={true} id="test-modal" isOpen={true}>
 				<XUIModalHeader/>
 				<XUIModalBody/>
 				<XUIModalFooter/>
@@ -178,4 +201,78 @@ describe('XUIModal', () => {
 		expect(domNode.is('form')).toBeTruthy();
 	});
 
+	it('should not render as a form by default, or if isForm is set to false', () => {
+		const component = mount(
+			<XUIModal isUsingPortal={false} id="test-modal" >
+				<XUIModalHeader/>
+				<XUIModalBody/>
+				<XUIModalFooter/>
+			</XUIModal>
+		);
+
+		const domNode = component.find('.xui-modal');
+		expect(domNode.is('form')).toBeFalsy();
+
+		const componentSet = mount(
+			<XUIModal isUsingPortal={false} isForm={false} id="test-modal">
+				<XUIModalHeader/>
+				<XUIModalBody/>
+				<XUIModalFooter/>
+			</XUIModal>
+		);
+
+		const secondDomNode = componentSet.find('.xui-modal');
+		expect(secondDomNode.is('form')).toBeFalsy();
+	});
+
+	it('should render an automation id when a qaHook is passed in', () => {
+		const automationid = renderer.create(<XUIModal
+			isUsingPortal={false}
+			qaHook="xui-modal"
+			onClose={NOOP}
+		>
+			<div>test</div>
+		</XUIModal>);
+
+		expect(automationid).toMatchSnapshot();
+	});
+
+	it('should lock the scroll of a window when mounted and unlocked when unmounted', () => {
+		const modalMounted = shallow(<XUIModal
+			isUsingPortal={false}
+			onClose={NOOP}
+			isOpen={true}
+		>
+			<div>test</div>
+		</XUIModal>);
+
+		expect(modalMounted.instance()._isScrollLocked).toBeTruthy();
+
+		modalMounted.setProps({isOpen: false});
+
+		expect(modalMounted.instance()._isScrollLocked).toBeFalsy();
+	});
+
+	it('should render closed by default', () => {
+		const modalMounted = shallow(<XUIModal
+			isUsingPortal={false}
+			onClose={NOOP}
+		>
+			<div>test</div>
+		</XUIModal>);
+
+		expect(modalMounted.hasClass('xui-mask-is-active')).toBeFalsy();
+	});
+
+	it('should render open when isOpen is set to true', () => {
+		const modalMounted = shallow(<XUIModal
+			isUsingPortal={false}
+			onClose={NOOP}
+			isOpen={true}
+		>
+			<div>test</div>
+		</XUIModal>);
+
+		expect(modalMounted.hasClass('xui-mask-is-active')).toBeTruthy();
+	});
 });